list the parts of each route
Now the parts of the route are actually picked up by Handlebars when traversing the JSON. Length and duration are listed but no actual styling or additional info is added.

Next step would be to format the routes in a meaningful way and to insert other data on each part such as line number and type.