fetch addresses when editing options
A call to the Reittiopas API is made when saving changes to the Home Address (nothing is actually saved as of yet). If there are multiple matches, they are listed in a SimpleDialog (http://dev.jtsage.com/) dialog.

Next step: do something with the address choice. Name and coordinates need to be saved to local storage and the function applied to all address fields.